backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai>=1.30.0
boto3==1.34.34
chromadb==0.4.22
aiosqlite==0.19.0
pypdf==4.0.1
sentence-transformers==2.3.1
sounddevice==0.4.6
amazon-transcribe==0.6.2
pytest-asyncio==1.3.0
hypothesis==6.92.0
